+++ SpaceXDashboard.Server/Entities/RocketLaunch.cs
@@ -1,0 +1,6 @@
+﻿namespace SpaceXDashboard.Server.Entities
+{
+    public class RocketLaunch
+    {
+    }
+}