+++ spacexdashboard.client/vite.config.js
@@ -47,7 +47,7 @@ export default defineConfig({
     },
     server: {
         proxy: {
-            '^/weatherforecast': {
+            '^/rocketlaunches': {
                 target,
                 secure: false
             }